requirements.txt
google-genai==1.48.0
openai==2.7.1
python-dotenv==1.2.1
streamlit==1.51.0